+++ x/wasm/ibctesting/chain.go
@@ -335,12 +335,12 @@ func (chain *TestChain) SendMsgs(msgs ...sdk.Msg) (*sdk.Result, error) {
 
 	chain.Coordinator.IncrementTime()
 
-	chain.captureIBCEvents(r)
+	chain.CaptureIBCEvents(r)
 
 	return r, nil
 }
 
-func (chain *TestChain) captureIBCEvents(r *sdk.Result) {
+func (chain *TestChain) CaptureIBCEvents(r *sdk.Result) {
 	toSend := getSendPackets(r.Events)
 	if len(toSend) > 0 {
 		// Keep a queue on the chain that we can relay in tests