+++ x/wasm/keeper/gas_register.go
@@ -68,7 +68,7 @@ func DefaultPerByteUncompressCost() wasmvmtypes.UFraction {
 
 // GasRegister abstract source for gas costs
 type GasRegister interface {
-	// NewContractInstanceCosts costs to crate a new contract instance from code
+	// NewContractInstanceCosts costs to create a new contract instance from code
 	NewContractInstanceCosts(pinned bool, msgLen int) sdk.Gas
 	// CompileCosts costs to persist and "compile" a new wasm contract
 	CompileCosts(byteLength int) sdk.Gas
@@ -147,7 +147,7 @@ func NewWasmGasRegister(c WasmGasRegisterConfig) WasmGasRegister {
 	}
 }
 
-// NewContractInstanceCosts costs to crate a new contract instance from code
+// NewContractInstanceCosts costs to create a new contract instance from code
 func (g WasmGasRegister) NewContractInstanceCosts(pinned bool, msgLen int) storetypes.Gas {
 	return g.InstantiateContractCosts(pinned, msgLen)
 }