+++ x/wasm/simulation/genesis.go
@@ -16,7 +16,6 @@ func RandomizedGenState(simstate *module.SimulationState) {
 		Sequences: []types.Sequence{
 			{IDKey: types.KeyLastCodeID, Value: simstate.Rand.Uint64()},
 		},
-		GenMsgs: nil,
 	}
 
 	_, err := simstate.Cdc.MarshalJSON(&wasmGenesis)