+++ x/wasm/types/genesis.go
@@ -32,11 +32,7 @@ func (s GenesisState) ValidateBasic() error {
 			return sdkerrors.Wrapf(err, "sequence: %d", i)
 		}
 	}
-	for i := range s.GenMsgs {
-		if err := s.GenMsgs[i].ValidateBasic(); err != nil {
-			return sdkerrors.Wrapf(err, "gen message: %d", i)
-		}
-	}
+
 	return nil
 }
 
@@ -78,28 +74,6 @@ func (c Contract) ValidateBasic() error {
 		}
 	}
 	return nil
-}
-
-// AsMsg returns the underlying cosmos-sdk message instance. Null when can not be mapped to a known type.
-func (m GenesisState_GenMsgs) AsMsg() sdk.Msg {
-	if msg := m.GetStoreCode(); msg != nil {
-		return msg
-	}
-	if msg := m.GetInstantiateContract(); msg != nil {
-		return msg
-	}
-	if msg := m.GetExecuteContract(); msg != nil {
-		return msg
-	}
-	return nil
-}
-
-func (m GenesisState_GenMsgs) ValidateBasic() error {
-	msg := m.AsMsg()
-	if msg == nil {
-		return sdkerrors.Wrapf(sdkerrors.ErrInvalidType, "unknown message")
-	}
-	return msg.ValidateBasic()
 }
 
 // ValidateGenesis performs basic validation of supply genesis data returning an