+++ x/wasm/types/test_fixtures.go
@@ -36,11 +36,7 @@ func GenesisFixture(mutators ...func(*GenesisState)) GenesisState {
 			Value: uint64(i),
 		}
 	}
-	fixture.GenMsgs = []GenesisState_GenMsgs{
-		{Sum: &GenesisState_GenMsgs_StoreCode{StoreCode: MsgStoreCodeFixture()}},
-		{Sum: &GenesisState_GenMsgs_InstantiateContract{InstantiateContract: MsgInstantiateContractFixture()}},
-		{Sum: &GenesisState_GenMsgs_ExecuteContract{ExecuteContract: MsgExecuteContractFixture()}},
-	}
+
 	for _, m := range mutators {
 		m(&fixture)
 	}
@@ -250,6 +246,46 @@ func InstantiateContractProposalFixture(mutators ...func(p *InstantiateContractP
 		Label:       "testing",
 		Msg:         initMsgBz,
 		Funds:       nil,
+	}
+
+	for _, m := range mutators {
+		m(p)
+	}
+	return p
+}
+
+func InstantiateContract2ProposalFixture(mutators ...func(p *InstantiateContract2Proposal)) *InstantiateContract2Proposal {
+	var (
+		anyValidAddress sdk.AccAddress = bytes.Repeat([]byte{0x1}, ContractAddrLen)
+
+		initMsg = struct {
+			Verifier    sdk.AccAddress `json:"verifier"`
+			Beneficiary sdk.AccAddress `json:"beneficiary"`
+		}{
+			Verifier:    anyValidAddress,
+			Beneficiary: anyValidAddress,
+		}
+	)
+	const (
+		anyAddress = "cosmos1qyqszqgpqyqszqgpqyqszqgpqyqszqgpqyqszqgpqyqszqgpqyqs2m6sx4"
+		mySalt     = "myDefaultSalt"
+	)
+
+	initMsgBz, err := json.Marshal(initMsg)
+	if err != nil {
+		panic(err)
+	}
+	p := &InstantiateContract2Proposal{
+		Title:       "Foo",
+		Description: "Bar",
+		RunAs:       anyAddress,
+		Admin:       anyAddress,
+		CodeID:      1,
+		Label:       "testing",
+		Msg:         initMsgBz,
+		Funds:       nil,
+		Salt:        []byte(mySalt),
+		FixMsg:      false,
 	}
 
 	for _, m := range mutators {