+++ x/wasm/types/types.go
@@ -396,3 +396,14 @@ func isSubset(super, sub []string) bool {
 	}
 	return matches == len(sub)
 }
+
+// AllAuthorizedAddresses returns the list of authorized addresses. Can be empty.
+func (a AccessConfig) AllAuthorizedAddresses() []string {
+	switch a.Permission {
+	case AccessTypeAnyOfAddresses:
+		return a.Addresses
+	case AccessTypeOnlyAddress:
+		return []string{a.Address}
+	}
+	return []string{}
+}